requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
cloudinary==1.41.0
Django==5.1
django-paypal==2.1
gunicorn==23.0.0
idna==3.8
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==10.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
